chore(releases): moving rest of bundles to releases
