+++ packages/sanity/src/core/index.ts
@@ -1,4 +1,3 @@
-export {getDocumentIsInPerspective, LATEST} from './bundles'
 export * from './changeIndicators'
 export {
   CommentInput,
@@ -22,6 +21,7 @@ export * from './hooks'
 export * from './i18n'
 export * from './presence'
 export * from './preview'
+export {getDocumentIsInPerspective, LATEST} from './releases'
 export {ReleaseActions, ReleaseBadge, ReleasesMenu} from './releases/components'
 export {useDocumentVersions, usePerspective} from './releases/hooks'
 export * from './scheduledPublishing'
