+++ packages/sanity/src/core/releases/hooks/usePerspective.tsx
@@ -1,6 +1,6 @@
 import {useRouter} from 'sanity/router'
 
-import {LATEST} from '../../bundles'
+import {LATEST} from '../../releases'
 import {useBundles} from '../../store/bundles'
 import {type BundleDocument} from '../../store/bundles/types'
