+++ packages/sanity/src/core/releases/index.ts
@@ -1,2 +1,4 @@
+export * from './components'
+export * from './hooks'
 export * from './util/const'
 export * from './util/util'
