+++ packages/sanity/src/core/releases/tool/components/ReleasePublishAllButton/ReleasePublishAllButton.tsx
@@ -5,9 +5,9 @@ import {useCallback, useMemo, useState} from 'react'
 import {type BundleDocument} from 'sanity'
 
 import {Button, Dialog} from '../../../../../ui-components'
-import {PublishedRelease} from '../../../../bundles/__telemetry__/releases.telemetry'
 import {Translate, useTranslation} from '../../../../i18n'
 import {useBundleOperations} from '../../../../store/bundles/useBundleOperations'
+import {PublishedRelease} from '../../../__telemetry__/releases.telemetry'
 import {releasesLocaleNamespace} from '../../../i18n'
 import {type DocumentInBundleResult} from '../../../tool/detail/useBundleDocuments'
 import {useObserveDocumentRevisions} from './useObserveDocumentRevisions'
